+++ modAionImpl/src/org/aion/zero/impl/config/dynamic/InFlightConfigChangeException.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.config.dynamic2;
+package org.aion.zero.impl.config.dynamic;
 
 public class InFlightConfigChangeException extends Exception {
     public InFlightConfigChangeException(String message) {
