+++ modAionImpl/src/org/aion/zero/impl/config/dynamic/InFlightConfigChangeNotAllowedException.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.config.dynamic2;
+package org.aion.zero.impl.config.dynamic;
 
 public class InFlightConfigChangeNotAllowedException extends InFlightConfigChangeException {
     public InFlightConfigChangeNotAllowedException(String message) {
