+++ modAionImpl/src/org/aion/zero/impl/config/dynamic/RollbackException.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.config.dynamic2;
+package org.aion.zero.impl.config.dynamic;
 
 import java.util.LinkedList;
 import java.util.List;
