+++ modAionImpl/test/org/aion/zero/impl/config/dynamic/InFlightConfigReceiverTest.java
@@ -1,4 +1,4 @@
-package org.aion.mcf.config.dynamic2;
+package org.aion.zero.impl.config.dynamic;
 
 import org.aion.mcf.config.Cfg;
 import org.apache.commons.lang3.tuple.ImmutablePair;
