+++ modBoot/src/org/aion/Aion.java
@@ -28,7 +28,6 @@ import static org.aion.zero.impl.Version.KERNEL_VERSION;
 
 import java.lang.management.ManagementFactory;
 import java.util.ServiceLoader;
-import java.util.concurrent.TimeUnit;
 
 import org.aion.api.server.http.NanoServer;
 import org.aion.api.server.pb.ApiAion0;
@@ -41,14 +40,10 @@ import org.aion.evtmgr.EventMgrModule;
 import org.aion.log.AionLoggerFactory;
 import org.aion.log.LogEnum;
 import org.aion.mcf.config.CfgApiRpc;
-import org.aion.mcf.config.applier.MiningApplier;
-import org.aion.mcf.config.dynamic2.DynamicConfigKeyRegistry;
-import org.aion.mcf.config.dynamic2.InFlightConfigReceiver;
-import org.aion.mcf.config.dynamic2.InFlightConfigReceiverMBean;
+import org.aion.zero.impl.config.dynamic.DynamicConfigKeyRegistry;
+import org.aion.zero.impl.config.dynamic.InFlightConfigReceiver;
 import org.aion.mcf.mine.IMineRunner;
-import org.aion.zero.impl.AionHub;
 import org.aion.zero.impl.blockchain.AionFactory;
-import org.aion.zero.impl.blockchain.AionImpl;
 import org.aion.zero.impl.blockchain.IAionChain;
 import org.aion.zero.impl.cli.Cli;
 import org.aion.zero.impl.config.CfgAion;
@@ -57,14 +52,9 @@ import org.slf4j.Logger;
 import javax.management.InstanceAlreadyExistsException;
 import javax.management.MBeanRegistrationException;
 import javax.management.MBeanServer;
-import javax.management.MBeanServerConnection;
-import javax.management.MBeanServerInvocationHandler;
 import javax.management.MalformedObjectNameException;
 import javax.management.NotCompliantMBeanException;
 import javax.management.ObjectName;
-import javax.management.remote.JMXConnector;
-import javax.management.remote.JMXConnectorFactory;
-import javax.management.remote.JMXServiceURL;
 
 public class Aion {
