+++ modGui/src/org/aion/gui/model/ConfigManipulator.java
@@ -6,10 +6,10 @@ import com.google.common.io.CharSource;
 import com.google.common.io.Files;
 import org.aion.log.AionLoggerFactory;
 import org.aion.mcf.config.Cfg;
-import org.aion.mcf.config.dynamic2.ConfigProposalResult;
-import org.aion.mcf.config.dynamic2.InFlightConfigReceiver;
-import org.aion.mcf.config.dynamic2.InFlightConfigReceiverMBean;
-import org.aion.mcf.config.dynamic2.RollbackException;
+import org.aion.zero.impl.config.dynamic.ConfigProposalResult;
+import org.aion.zero.impl.config.dynamic.InFlightConfigReceiver;
+import org.aion.zero.impl.config.dynamic.InFlightConfigReceiverMBean;
+import org.aion.zero.impl.config.dynamic.RollbackException;
 import org.aion.os.KernelLauncher;
 import org.aion.zero.impl.config.CfgAion;
 import org.slf4j.Logger;
